backend/requirements.txt
flask
opencv-python-headless
numpy
joblib
pillow
rembg
scikit-image
scikit-learn
gunicorn